requirements.txt
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
binaryornot==0.4.4
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cookiecutter==2.6.0
decorator==5.2.1
distlib==0.3.9
dj-database-url==1.3.0
Django==5.1.5
django-browser-reload==1.17.0
django-ckeditor==6.7.2
django-cors-headers==4.7.0
django-js-asset==3.1.2
django-jsoneditor==0.2.4
django-tailwind==3.8.0
djangorestframework==3.16.0
executing==2.2.0
filelock==3.18.0
fsspec==2025.3.2
git-filter-repo==2.47.0
gunicorn==23.0.0
huggingface-hub==0.31.1
idna==3.10
ipdb==0.13.13
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.5
joblib==1.5.0
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pure_eval==0.2.3
Pygments==2.19.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.4.2
requests==2.32.3
rich==13.9.4
safetensors==0.5.3
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.13.3
text-unidecode==1.3
tokenizers==0.21.1
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.3.0
virtualenv==20.29.3
wcwidth==0.2.13
wheel==0.44.0
whitenoise==6.9.0
